Line-of-Sight works!  Needs considerable clean-up!
The current implementation does not always produce perfect results
because it uses the center of two endpoint cells as the basis for the DDA
calculations.

Also: This code should be cleaned up to provide better separation between
"grid" or map coordinates and screen coordinates.  Animations, for
example, are generally going to require screen coordinates.  However,
nearly everything else (movement, collision detection, item placement,
etc.) will be based on map coordinates (i.e. in terms of cells, not
pixels.)

Still, this is exciting progress!

Signed-off-by: Atom Macumber <amacumber@eagleproductivity.com>