Add isTransparent method to Map, adjust DDAMap accordingly.
Currently, a cell's transparency is determined by whether or not it is
passable.  This is obviously not a reliable indicator (i.e. deep water)
but since the test map does not contain any data for transparency and
the concept of monster hostility isn't implemented yet, this method will
suffice for the time being.

Signed-off-by: Atom Macumber <amacumber@eagleproductivity.com>