Huge code restructure.  Too much to mention...
Same functionality as before but the code has been moved around quite a bit so as to
increase flexibility and ease of use.  There's still a lot of really gross stuff sitting out
there but this is a solid start!

TODO items have been left in so I can remember what I messed up.