Greatly improve viewport performance.
Movement was very slow/choppy, especially if the viewport was following
the player.  This appears to be due to the high number of entities that
were being created to represent each tile/cell of the map and the
calculations that Crafty was doing to determine whether each cell was
within the viewport's boundaries.

By representing the cells with non-entities (in this case a Function
object), the cell/map data can be represented without all of the overhead
introduced by Crafty.

Signed-off-by: suicidepills <atom.mac@gmail.com>